WCM-43: Make room in zeit.cms.testing so we can add SQL-specific code more easily
